Make sessions an optional part of the web application
